refactor: remove unused includes and forward declarations from headers
