+++ src/evo/deterministicmns.h
@@ -29,6 +29,7 @@
 class CBlock;
 class CBlockIndex;
 class CChainState;
+class CCoinsViewCache;
 class CConnman;
 class CEvoDB;
 class TxValidationState;
