+++ src/evo/simplifiedmns.h
@@ -6,17 +6,21 @@
 #define BITCOIN_EVO_SIMPLIFIEDMNS_H
 
 #include <bls/bls.h>
-#include <evo/deterministicmns.h>
 #include <evo/dmn_types.h>
 #include <merkleblock.h>
 #include <netaddress.h>
 #include <pubkey.h>
+#include <sync.h>
+#include <threadsafety.h>
 
 class UniValue;
 class CBlockIndex;
-class CDeterministicMNList;
 class CDeterministicMN;
+class CDeterministicMNList;
+class CDeterministicMNManager;
 class ChainstateManager;
+
+extern RecursiveMutex cs_main;
 
 namespace llmq {
 class CFinalCommitment;
